Clean up the pk414 package a bit
We don't want a "laag" kernel. We want a generic standard kernel that
is broadly usable across many configurations and systems.
We explicitly do not want a "kernel of the day" for each random SOC;
we can and will reuse kernels.

So rename the laag sub kernel, and harmonize the configuration with
the normal Clear Linux kernel configurations